fix(producer): use an alias for replying sync messages and check if pid is alive
When one restarts the Erlang VM and reads PIDs from it back, when
attempting to send messages (replies) to those PIDs from an old
incarnation of the VM, some messages are printed:

```
(emqx@127.0.0.1)1> 2022-10-10T10:29:37.452699-03:00 [error] Discarding message {#Ref<0.1037307208.3830448135.240372>,{ok,#{highest_sequence_id=>0,message_id=>#{ack_set=>[],entryId=>0,ledgerId=>75},producer_id=>1,sequence_id=>1}}} from <0.2697.0> to <0.2869.0> in an old incarnation (3) of this node (1)
```

Here, we check whether such PIDs are still alive.  If not, we replace
the `From` with `undefined`, as if the request was async.