fix(producer): always lookup topic during startup
If the network is partitioned when a producer is connected to Pulsar,
but Pulsar itself doesn't go down, then there's no need to lookup the
topic to which the producer will produce (as this is done by
`pulsar_client`) prior to starting the producer the first time.

If, however, the Pulsar broker goes down and is later brought back up,
the producer MUST lookup the topic again before sending any messages.
Otherwise, Pulsar will reply with an error of the form:

```
Namespace bundle for topic (persistent://public/default/some-topic)
not served by this instance. Please redo the lookup. Request is
denied: namespace=public/default
```

Reference: https://pulsar.apache.org/docs/2.10.x/developing-binary-protocol/#topic-lookup