fix(producer): request topic lookup asynchronously
Making a sync call to `pulsar_client` to perform a
`CommandLookupTopic` might hang for a lot of time, especially if the
network/connection has high latency.  Messages to be sent that are
received by `pulsar_producer` during this period of time will be lost
if the producer happens to be killed before queuing then.

We make it async here to avoid blocking on such calls to
`pulsar_client'.