fix(client): avoid client reaching maximum restart intensity during init
If a client crashes and is later restarted by the supervisor while
Pulsar is having a bad day (or the connection to it is), then it might
reach maximum restart intensity and be removed from
`pulsar_client_sup`.

Here, we retry the connection a few times, with a back-off time
between attempts, before giving up to avoid this situation.